requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.16.1
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
narwhals==1.33.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai==1.70.0
openai-whisper==20240930
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.20.0
protobuf==5.29.4
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
python-dotenv==1.0.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.7.0
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
zipp==3.20.2
python-docx==1.1.2
PyPDF2==3.0.1